Adjusted Upload/Download & Error
Fixed the upload and download functions that can be copy and pasted into the server so that they can upload files and download files to and from each other. The upload file will take the written file path and send the contents to the server, the server will write out the contents of the uploaded file to another file. They both worked individually. Also I believe I fixed the error in Running Processes.